Dedup output mode handling
